Made new models for data mig dependant on linapp.0003.
Previously, they were dependant on linapp.0001, so they wouldn't go back
even when moving to linapp.0001. Now they are dependant on 0003, so
moving to 0001 indeed reverts to pre-split state.
Also split workflows.0001 into two, because one has to run before
linapp.0003, and the other has to run after parts.0001 (which is now
dependent on linapp.0003).