Added planning_version field to new TEs.
This field will help us tell apart TEs that where planned by different
planners. Currently, we mark two versions of plannings: those done in
the no-tail era and those done in the "pcr1", with-tail era.
The values are left opaque, to be migrated into a more informative
scheme in the future.